analytics: Extract common /activity views functions.
This will be helpful for further splitting the giant files here.
